# v2/src/verification/api/openapi/specification.yaml
openapi: 3.0.0
info:
  title: Cortex Agent Verification API
  version: 1.0.0
  description: |
    Comprehensive verification API for truth monitoring and agent verification in Cortex Agent.
    
    This API provides:
    - REST endpoints for verification status and management
    - WebSocket connections for real-time truth monitoring
    - GraphQL interface for complex queries
    - Client SDKs for seamless agent integration
    
    ## Authentication
    All endpoints require API key authentication via the `X-API-Key` header.
    
    ## Rate Limiting
    API requests are rate limited per agent:
    - Default: 60 requests per minute
    - Burst capacity: 100 requests
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format.
    
  contact:
    name: Cortex Agent API Support
    url: https://github.com/ruvnet/cortex-agent
    email: support@cortex-agent.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cortex-agent.dev/v1
    description: Production server
  - url: https://staging-api.cortex-agent.dev/v1
    description: Staging server
  - url: http://localhost:3000/api/verification
    description: Local development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
  /status:
    get:
      summary: Get verification system status
      description: Returns overall system status, metrics, and recent events
      operationId: getVerificationStatus
      tags:
        - Status
      parameters:
        - name: limit
          in: query
          description: Number of recent events to include
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: System status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                success: true
                data:
                  status: operational
                  metrics:
                    total_verifications: 15420
                    successful_verifications: 14891
                    failed_verifications: 529
                    pending_verifications: 45
                    success_rate: 96.57
                    verification_rate: 127.3
                    average_confidence: 0.847
                  recent_events:
                    - type: verification_complete
                      verification_id: ver_123abc
                      timestamp: 1704067200000
                  active_verifications:
                    - id: ver_pending_1
                      source: agent_alpha
                      target: data_source_beta
                      status: pending
        '500':
          $ref: '#/components/responses/InternalServerError'
  /verify:
    post:
      summary: Create new verification
      description: Submit a new verification request for processing
      operationId: createVerification
      tags:
        - Verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVerificationRequest'
            example:
              source: agent_researcher_1
              target: scientific_paper_xyz
              metadata:
                paper_title: "Quantum Computing Advances"
                confidence_required: 0.9
              priority: high
              timeout: 45000
      responses:
        '201':
          description: Verification created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResponse'
              example:
                success: true
                data:
                  id: ver_abc123def
                  timestamp: 1704067200000
                  status: pending
                  confidence: 0
                  source: agent_researcher_1
                  target: scientific_paper_xyz
                  priority: high
                  created_at: 1704067200000
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /verify/{id}:
    get:
      summary: Get verification by ID
      description: Retrieve details of a specific verification
      operationId: getVerification
      tags:
        - Verification
      parameters:
        - name: id
          in: path
          required: true
          description: Verification ID
          schema:
            type: string
            pattern: '^ver_[a-zA-Z0-9]+$'
          example: ver_abc123def
      responses:
        '200':
          description: Verification retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /query:
    get:
      summary: Query verifications
      description: Search and filter verifications with advanced criteria
      operationId: queryVerifications
      tags:
        - Query
      parameters:
        - name: sources
          in: query
          description: Filter by source agents (comma-separated)
          schema:
            type: string
          example: agent_1,agent_2
        - name: targets
          in: query
          description: Filter by target resources (comma-separated)
          schema:
            type: string
          example: resource_a,resource_b
        - name: status
          in: query
          description: Filter by verification status (comma-separated)
          schema:
            type: string
            enum: [pending, verified, failed, unknown]
          example: verified,failed
        - name: confidence_min
          in: query
          description: Minimum confidence score
          schema:
            type: number
            minimum: 0
            maximum: 1
          example: 0.8
        - name: confidence_max
          in: query
          description: Maximum confidence score
          schema:
            type: number
            minimum: 0
            maximum: 1
          example: 1.0
        - name: timestamp_start
          in: query
          description: Start timestamp (Unix milliseconds)
          schema:
            type: integer
          example: 1704060000000
        - name: timestamp_end
          in: query
          description: End timestamp (Unix milliseconds)
          schema:
            type: integer
          example: 1704070000000
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
          example: 100
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /batch:
    post:
      summary: Create batch verification
      description: Submit multiple verification requests as a single batch
      operationId: createBatchVerification
      tags:
        - Batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
            example:
              items:
                - source: agent_1
                  target: resource_a
                  priority: normal
                - source: agent_1
                  target: resource_b
                  priority: high
              metadata:
                batch_name: "Weekly verification batch"
      responses:
        '201':
          description: Batch created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /batch/{id}:
    get:
      summary: Get batch status
      description: Retrieve the status and progress of a verification batch
      operationId: getBatchStatus
      tags:
        - Batch
      parameters:
        - name: id
          in: path
          required: true
          description: Batch ID
          schema:
            type: string
            pattern: '^batch_[a-zA-Z0-9]+$'
          example: batch_xyz789
      responses:
        '200':
          description: Batch status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    CreateVerificationRequest:
      type: object
      required:
        - source
        - target
      properties:
        source:
          type: string
          description: Source agent or system identifier
          example: agent_researcher_1
        target:
          type: string
          description: Target resource or data identifier
          example: scientific_paper_xyz
        metadata:
          type: object
          additionalProperties: true
          description: Additional context and parameters
          example:
            paper_title: "Quantum Computing Advances"
            confidence_required: 0.9
        priority:
          $ref: '#/components/schemas/Priority'
        timeout:
          type: integer
          minimum: 1000
          maximum: 300000
          description: Timeout in milliseconds
          default: 30000
          example: 45000
    CreateBatchRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/CreateVerificationRequest'
          description: Array of verification requests
        metadata:
          type: object
          additionalProperties: true
          description: Batch-level metadata
          example:
            batch_name: "Weekly verification batch"
...
    StatusResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                status:
                  type: string
                  enum: [operational, degraded, busy, maintenance]
                  example: operational
                metrics:
                  $ref: '#/components/schemas/VerificationMetrics'
                recent_events:
                  type: array
                  items:
                    $ref: '#/components/schemas/TruthMonitoringEvent'
                active_verifications:
                  type: array
                  items:
                    $ref: '#/components/schemas/Verification'
    VerificationResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Verification'
    BatchResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/VerificationBatch'
    QueryResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                results:
                  type: array
                  items:
                    $ref: '#/components/schemas/Verification'
                total:
                  type: integer
                  description: Total number of matching results
                  example: 1250
                pagination:
                  type: object
                  properties:
                    offset:
                      type: integer
                      example: 0
                    limit:
                      type: integer
                      example: 50
                    has_more:
                      type: boolean
                      example: true
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          required:
            - error
          properties:
            success:
              type: boolean
              enum: [false]
            data:
              type: "null"
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INVALID_REQUEST
              message: "Both source and target are required"
            metadata:
              timestamp: 1704067200000
              request_id: req_abc123
              version: "1.0.0"
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: "Invalid API key"
            metadata:
              timestamp: 1704067200000
              request_id: req_abc123
              version: "1.0.0"
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VERIFICATION_NOT_FOUND
              message: "Verification ver_123 not found"
            metadata:
              timestamp: 1704067200000
              request_id: req_abc123
              version: "1.0.0"
    RateLimited:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: The number of allowed requests in the current period
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: The number of remaining requests in the current period
          schema:
            type: integer
        X-RateLimit-Reset:
          description: The time at which the current rate limit window resets in UTC epoch seconds
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: "Rate limit exceeded. Try again in 60 seconds."
            metadata:
              timestamp: 1704067200000
              request_id: req_abc123
              version: "1.0.0"
    InternalServerError:
      description: Internal Server Error - Something went wrong on the server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: "An unexpected error occurred"
            metadata:
              timestamp: 1704067200000
              request_id: req_abc123
              version: "1.0.0"
...